+++ data/src/main/java/com/darotpeacedude/data/local/Result.kt
@@ -27,7 +27,7 @@ data class Result (
     val title: String,
     val video: Boolean,
     @SerializedName("vote_average")
-    val voteAverage: Long,
+    val voteAverage: Double,
     @SerializedName("vote_count")
     val voteCount: Long
 ):Serializable, Parent
